feat: improve update subscription
