+++ src/AeFinder.Application/BlockScan/BlockScanAppService.cs
@@ -1,14 +1,11 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text.Json;
 using System.Threading.Tasks;
 using AeFinder.App.Deploy;
 using AeFinder.Grains;
 using AeFinder.Grains.Grain.BlockPush;
 using AeFinder.Grains.Grain.BlockStates;
 using AeFinder.Grains.Grain.Subscriptions;
-using AeFinder.Studio;
 using Microsoft.Extensions.Logging;
 using Orleans;
 using Volo.Abp;
@@ -28,322 +25,7 @@ public class BlockScanAppService : AeFinderAppService, IBlockScanAppService
         _clusterClient = clusterClient;
         _kubernetesAppManager = kubernetesAppManager;
     }
-
-    public async Task<string> AddSubscriptionAsync(string appId, SubscriptionManifestDto manifestDto, byte[] dll = null)
-    {
-        var dto = await AddSubscriptionV2Async(appId, manifestDto, dll);
-        return dto.NewVersion;
-    }
-
-    public async Task<AddSubscriptionDto> AddSubscriptionV2Async(string appId, SubscriptionManifestDto manifestDto, byte[] dll = null)
-    {
-        Logger.LogInformation("ScanApp: {clientId} submit subscription: {subscription}", appId,
-            JsonSerializer.Serialize(manifestDto));
-
-        var subscription = ObjectMapper.Map<SubscriptionManifestDto, SubscriptionManifest>(manifestDto);
-        var client = _clusterClient.GetGrain<IAppSubscriptionGrain>(GrainIdHelper.GenerateAppSubscriptionGrainId(appId));
-        // TODO: Use the code from developer.
-        if (dll == null)
-        {
-            return null;
-        }
-
-        return await client.AddSubscriptionV2Async(subscription, dll);
-    }
-
-    public async Task UpdateSubscriptionInfoAsync(string appId, string version,
-        SubscriptionManifestDto manifestDto)
-    {
-        Logger.LogInformation($"Client: {appId} version: {version} update subscribe: {JsonSerializer.Serialize(manifestDto)}");
-        var appSubscriptionGrain = _clusterClient.GetGrain<IAppSubscriptionGrain>(GrainIdHelper.GenerateAppSubscriptionGrainId(appId));
-
-        var subscription = ObjectMapper.Map<SubscriptionManifestDto, SubscriptionManifest>(manifestDto);
-        var currentSubscriptionInfos = await appSubscriptionGrain.GetSubscriptionAsync(version);
-        
-        //Check input subscription info if is valid
-        CheckInputSubscriptionInfoIsValid(subscription.SubscriptionItems, currentSubscriptionInfos.SubscriptionItems);
-        CheckInputSubscriptionInfoIsDuplicateOrMissing(subscription.SubscriptionItems,currentSubscriptionInfos.SubscriptionItems);
-
-        await appSubscriptionGrain.UpdateSubscriptionAsync(version, subscription);
-        await _kubernetesAppManager.RestartAppAsync(appId, version);
-    }
-
-    private void CheckInputSubscriptionInfoIsValid(List<Subscription> subscriptionInfos,
-        List<Subscription> currentSubscriptionInfos)
-    {
-        foreach (var subscriptionInfo in subscriptionInfos)
-        {
-            var currentSubscriptionInfoForCheckChainId = currentSubscriptionInfos.FindAll(i =>
-                (i.ChainId == subscriptionInfo.ChainId));
-            if (currentSubscriptionInfoForCheckChainId == null || currentSubscriptionInfoForCheckChainId.Count == 0)
-            {
-                var errorMessage = $"Invalid chain id {subscriptionInfo.ChainId}, can not add new chain";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-
-            var currentSubscriptionInfo = currentSubscriptionInfoForCheckChainId.FirstOrDefault();
-            if ((currentSubscriptionInfo.TransactionConditions == null ||
-                 currentSubscriptionInfo.TransactionConditions.Count == 0) &&
-                (subscriptionInfo.TransactionConditions != null && subscriptionInfo.TransactionConditions.Count > 0))
-            {
-                var errorMessage = $"Can not add transactionConditions in chain {subscriptionInfo.ChainId}";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-            
-            if ((currentSubscriptionInfo.LogEventConditions == null ||
-                 currentSubscriptionInfo.LogEventConditions.Count == 0) &&
-                (subscriptionInfo.LogEventConditions != null && subscriptionInfo.LogEventConditions.Count > 0))
-            {
-                var errorMessage = $"Can not add logEventConditions in chain {subscriptionInfo.ChainId}";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-            
-            var currentSubscriptionInfoForCheckStartBlockNumber = currentSubscriptionInfoForCheckChainId.FindAll(i =>
-                i.StartBlockNumber == subscriptionInfo.StartBlockNumber);
-            if (currentSubscriptionInfoForCheckStartBlockNumber == null || currentSubscriptionInfoForCheckStartBlockNumber.Count == 0)
-            {
-                var errorMessage =
-                    $"Invalid start block number {subscriptionInfo.StartBlockNumber}, can not update start block number in chain {subscriptionInfo.ChainId}";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-            
-            var currentSubscriptionForCheckIsOnlyConfirmed = currentSubscriptionInfoForCheckStartBlockNumber.FindAll(i =>
-                i.OnlyConfirmed == subscriptionInfo.OnlyConfirmed);
-            if (currentSubscriptionForCheckIsOnlyConfirmed == null || currentSubscriptionForCheckIsOnlyConfirmed.Count == 0)
-            {
-                var errorMessage =
-                    $"Invalid only confirmed block {subscriptionInfo.OnlyConfirmed}, can not update only confirmed block in chain {subscriptionInfo.ChainId}";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-            
-        }
-    }
-
-    private void CheckInputSubscriptionInfoIsDuplicateOrMissing(List<Subscription> subscriptionInfos,
-        List<Subscription> currentSubscriptionInfos)
-    {
-        foreach (var currentSubscriptionInfo in currentSubscriptionInfos)
-        {
-            var subscriptionInfoForCheckDuplicate= subscriptionInfos.FindAll(i =>
-                (i.ChainId == currentSubscriptionInfo.ChainId));
-            if (subscriptionInfoForCheckDuplicate != null && subscriptionInfoForCheckDuplicate.Count > 1)
-            {
-                var errorMessage =
-                    $"Duplicate subscribe information in chain {currentSubscriptionInfo.ChainId}";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-            
-            var subscriptionInfoForCheck = subscriptionInfoForCheckDuplicate.FirstOrDefault(i =>
-                (i.StartBlockNumber == currentSubscriptionInfo.StartBlockNumber && i.OnlyConfirmed == currentSubscriptionInfo.OnlyConfirmed));
-            if (subscriptionInfoForCheck == null)
-            {
-                var errorMessage =
-                    $"Can not modify StartBlockNumber or OnlyConfirmed of subscribe information in chain {currentSubscriptionInfo.ChainId}";
-                throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-            }
-            
-            if (currentSubscriptionInfo.TransactionConditions != null &&
-                currentSubscriptionInfo.TransactionConditions.Count > 0)
-            {
-                foreach (var transactionCondition in currentSubscriptionInfo.TransactionConditions)
-                {
-                    var currentTo = transactionCondition.To;
-                    var subscriptionInfoForCheckTransactionCondition =
-                        subscriptionInfoForCheck.TransactionConditions.FirstOrDefault(i =>
-                            (i.To == currentTo));
-                    if (subscriptionInfoForCheckTransactionCondition == null)
-                    {
-                        var errorMessage =
-                            $"Can not remove subscribed transaction condition of to address {currentTo} in chain {subscriptionInfoForCheck.ChainId}";
-                        throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-                    }
-
-                    foreach (var methodName in transactionCondition.MethodNames)
-                    {
-                        var inputMethodName=subscriptionInfoForCheckTransactionCondition.MethodNames.FirstOrDefault(i => i == methodName);
-                        if (inputMethodName.IsNullOrEmpty())
-                        {
-                            var errorMessage =
-                                $"Can not remove subscribed transaction condition of method name {methodName} in chain {subscriptionInfoForCheck.ChainId} to address {currentTo}";
-                            throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-                        }
-                    }
-                }
-            }
-            
-            if (currentSubscriptionInfo.LogEventConditions != null &&
-                currentSubscriptionInfo.LogEventConditions.Count > 0)
-            {
-                foreach (var logEventCondition in currentSubscriptionInfo.LogEventConditions)
-                {
-                    var currentContractAddress = logEventCondition.ContractAddress;
-                    var subscriptionInfoForCheckLogEventCondition =
-                        subscriptionInfoForCheck.LogEventConditions.FirstOrDefault(i =>
-                            (i.ContractAddress == currentContractAddress));
-                    if (subscriptionInfoForCheckLogEventCondition == null)
-                    {
-                        var errorMessage =
-                            $"Can not remove subscribe log event condition of contract address {currentContractAddress} in chain {subscriptionInfoForCheck.ChainId}";
-                        throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-                    }
-
-                    foreach (var eventName in logEventCondition.EventNames)
-                    {
-                        var inputEventName=subscriptionInfoForCheckLogEventCondition.EventNames.FirstOrDefault(i => i == eventName);
-                        if (inputEventName.IsNullOrEmpty())
-                        {
-                            var errorMessage =
-                                $"Can not remove subscribe log event condition of event name {eventName} in chain {subscriptionInfoForCheck.ChainId} contract address {currentContractAddress}";
-                            throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-                        }
-                    }
-                }
-            }
-            
-            
-            
-        }
-    }
-
-    // public async Task UpdateSubscriptionInfoAsync(string clientId, string version, List<SubscriptionInfo> subscriptionInfos)
-    // {
-    //     Logger.LogInformation($"Client: {clientId} update subscribe: {JsonSerializer.Serialize(subscriptionInfos)}");
-    //     
-    //     var client = _clusterClient.GetGrain<IClientGrain>(clientId);
-    //     
-    //     //Check if the subscription info is valid
-    //     var currentSubscriptionInfos = await client.GetSubscriptionAsync(version);
-    //
-    //     CheckInputSubscriptionInfoIsValid(subscriptionInfos, currentSubscriptionInfos);
-    //
-    //     CheckInputSubscriptionInfoIsDuplicateOrMissing(subscriptionInfos, currentSubscriptionInfos);
-    //     
-    //     //Update subscription info
-    //     await client.UpdateSubscriptionInfoAsync(version, subscriptionInfos);
-    //
-    //     foreach (var subscriptionInfo in subscriptionInfos)
-    //     {
-    //         var id = GrainIdHelper.GenerateGrainId(subscriptionInfo.ChainId, clientId, version, subscriptionInfo.FilterType);
-    //         var blockScanInfoGrain = _clusterClient.GetGrain<IBlockScanGrain>(id);
-    //         await blockScanInfoGrain.UpdateSubscriptionInfoAsync(subscriptionInfo);
-    //     }
-    // }
-
-    // private void CheckInputSubscriptionInfoIsValid(List<SubscriptionInfo> subscriptionInfos,
-    //     List<SubscriptionInfo> currentSubscriptionInfos)
-    // {
-    //     foreach (var subscriptionInfo in subscriptionInfos)
-    //     {
-    //         var subscriptionInfoForCheckChainId = currentSubscriptionInfos.FindAll(i =>
-    //             (i.ChainId == subscriptionInfo.ChainId));
-    //         if (subscriptionInfoForCheckChainId == null || subscriptionInfoForCheckChainId.Count == 0)
-    //         {
-    //             var errorMessage = $"Invalid chain id {subscriptionInfo.ChainId}, can not add new chain";
-    //             throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //         }
-    //
-    //         var subscriptionInfoForCheckFilterType = subscriptionInfoForCheckChainId.FindAll(i =>
-    //             i.FilterType == subscriptionInfo.FilterType);
-    //         if (subscriptionInfoForCheckFilterType == null || subscriptionInfoForCheckFilterType.Count == 0)
-    //         {
-    //             continue;
-    //         }
-    //
-    //         var subscriptionInfoForCheckStartBlockNumber = subscriptionInfoForCheckFilterType.FindAll(i =>
-    //             i.StartBlockNumber == subscriptionInfo.StartBlockNumber);
-    //         if (subscriptionInfoForCheckStartBlockNumber == null || subscriptionInfoForCheckStartBlockNumber.Count == 0)
-    //         {
-    //             var errorMessage =
-    //                 $"Invalid start block number {subscriptionInfo.StartBlockNumber}, can not update start block number in chain {subscriptionInfo.ChainId} filterType {subscriptionInfo.FilterType}";
-    //             throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //         }
-    //
-    //         var subscriptionInfoForCheckIsOnlyConfirmed = subscriptionInfoForCheckStartBlockNumber.FindAll(i =>
-    //             i.OnlyConfirmedBlock == subscriptionInfo.OnlyConfirmedBlock);
-    //         if (subscriptionInfoForCheckIsOnlyConfirmed == null || subscriptionInfoForCheckIsOnlyConfirmed.Count == 0)
-    //         {
-    //             var errorMessage =
-    //                 $"Invalid only confirmed block {subscriptionInfo.OnlyConfirmedBlock}, can not update only confirmed block in chain {subscriptionInfo.ChainId} filterType {subscriptionInfo.FilterType}";
-    //             throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //         }
-    //
-    //         var subscriptionInfoForCheckContract = subscriptionInfoForCheckIsOnlyConfirmed.FirstOrDefault();
-    //         if (subscriptionInfo.SubscribeEvents == null || subscriptionInfo.SubscribeEvents.Count == 0)
-    //         {
-    //             if (subscriptionInfoForCheckContract.SubscribeEvents != null &&
-    //                 subscriptionInfoForCheckContract.SubscribeEvents.Count > 0)
-    //             {
-    //                 var errorMessage =
-    //                     $"Can not empty subscribe contracts in chain {subscriptionInfo.ChainId} filterType {subscriptionInfo.FilterType}";
-    //                 throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //             }
-    //         }
-    //         else
-    //         {
-    //             if (subscriptionInfoForCheckContract.SubscribeEvents == null ||
-    //                 subscriptionInfoForCheckContract.SubscribeEvents.Count == 0)
-    //             {
-    //                 var errorMessage =
-    //                     $"Can not add subscribe contracts in chain {subscriptionInfo.ChainId} filterType {subscriptionInfo.FilterType}";
-    //                 throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //             }
-    //
-    //             foreach (var filterContractEventInput in subscriptionInfo.SubscribeEvents)
-    //             {
-    //                 var subscriptionInfoForCheckContractEvent =
-    //                     subscriptionInfoForCheckContract.SubscribeEvents.FirstOrDefault(i =>
-    //                         (i.ContractAddress == filterContractEventInput.ContractAddress));
-    //                 if (subscriptionInfoForCheckContractEvent == null)
-    //                 {
-    //                     var errorMessage =
-    //                         $"Can not add new subscribe contract {filterContractEventInput.ContractAddress} in chain {subscriptionInfo.ChainId} filterType {subscriptionInfo.FilterType}";
-    //                     throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //                 }
-    //             }
-    //         }
-    //     }
-    // }
-    //
-    // private void CheckInputSubscriptionInfoIsDuplicateOrMissing(List<SubscriptionInfo> subscriptionInfos,
-    //     List<SubscriptionInfo> currentSubscriptionInfos)
-    // {
-    //     foreach (var currentSubscriptionInfo in currentSubscriptionInfos)
-    //     {
-    //         var subscriptionInfoForCheckDuplicate= subscriptionInfos.FindAll(i =>
-    //             (i.ChainId == currentSubscriptionInfo.ChainId && i.FilterType == currentSubscriptionInfo.FilterType));
-    //         if (subscriptionInfoForCheckDuplicate != null && subscriptionInfoForCheckDuplicate.Count > 1)
-    //         {
-    //             var errorMessage =
-    //                 $"Duplicate subscribe information in chain {currentSubscriptionInfo.ChainId} filterType {currentSubscriptionInfo.FilterType}";
-    //             throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //         }
-    //         
-    //         var subscriptionInfoForCheck = subscriptionInfoForCheckDuplicate.FirstOrDefault(i =>
-    //             (i.StartBlockNumber == currentSubscriptionInfo.StartBlockNumber && i.OnlyConfirmedBlock == currentSubscriptionInfo.OnlyConfirmedBlock));
-    //         if (subscriptionInfoForCheck == null)
-    //         {
-    //             var errorMessage =
-    //                 $"Missing subscribe information in chain {currentSubscriptionInfo.ChainId} filterType {currentSubscriptionInfo.FilterType}";
-    //             throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //         }
-    //
-    //         if (currentSubscriptionInfo.SubscribeEvents != null && currentSubscriptionInfo.SubscribeEvents.Count > 0)
-    //         {
-    //             foreach (var currentSubscribeContract in currentSubscriptionInfo.SubscribeEvents)
-    //             {
-    //                 var subscriptionInfoForCheckContractEvent = subscriptionInfoForCheck.SubscribeEvents.FirstOrDefault(i =>
-    //                     (i.ContractAddress == currentSubscribeContract.ContractAddress));
-    //                 if (subscriptionInfoForCheckContractEvent == null)
-    //                 {
-    //                     var errorMessage =
-    //                         $"Can not remove subscribe contract {currentSubscribeContract.ContractAddress} in chain {currentSubscriptionInfo.ChainId} filterType {currentSubscriptionInfo.FilterType}";
-    //                     throw new UserFriendlyException("Invalid subscriptionInfo", details: errorMessage);
-    //                 }
-    //             }
-    //         }
-    //     }
-    // }
-
+    
     public async Task<List<Guid>> GetMessageStreamIdsAsync(string clientId, string version)
     {
         var client = _clusterClient.GetGrain<IAppSubscriptionGrain>(GrainIdHelper.GenerateAppSubscriptionGrainId(clientId));
