+++ src/AeFinder.Grains/Grain/Subscriptions/IAppSubscriptionGrain.cs
@@ -5,8 +5,7 @@ namespace AeFinder.Grains.Grain.Subscriptions;
 
 public interface IAppSubscriptionGrain : IGrainWithStringKey
 {
-    Task<string> AddSubscriptionAsync(SubscriptionManifest subscriptionManifest, byte[] code);
-    Task<AddSubscriptionDto> AddSubscriptionV2Async(SubscriptionManifest subscriptionManifest, byte[] code);
+    Task<AddSubscriptionDto> AddSubscriptionAsync(SubscriptionManifest subscriptionManifest, byte[] code);
     Task UpdateSubscriptionAsync(string version, SubscriptionManifest subscriptionManifest);
     Task UpdateCodeAsync(string version, byte[] code);
     Task<SubscriptionManifest> GetSubscriptionAsync(string version);
