+++ test/AeFinder.App.Tests/Handlers/SubscribedBlockHandlerTests.cs
@@ -48,7 +48,7 @@ public class SubscribedBlockHandlerTests : AeFinderAppTestBase
         var chainId = "AELF";
         var appGrain = Cluster.Client.GetGrain<IAppSubscriptionGrain>(GrainIdHelper.GenerateAppSubscriptionGrainId(_appInfoProvider.AppId));
         
-        var currentVersion = await appGrain.AddSubscriptionAsync(new SubscriptionManifest
+        var currentVersion = (await appGrain.AddSubscriptionAsync(new SubscriptionManifest
         {
             SubscriptionItems = new List<Subscription>
             {
@@ -58,7 +58,7 @@ public class SubscribedBlockHandlerTests : AeFinderAppTestBase
                     StartBlockNumber = 10000
                 }
             }
-        }, new byte[] { });
+        }, new byte[] { })).NewVersion;
         _appInfoProvider.SetVersion(currentVersion);
 
         var blockPusherInfoGrain = Cluster.Client.GetGrain<IBlockPusherInfoGrain>(
@@ -153,7 +153,7 @@ public class SubscribedBlockHandlerTests : AeFinderAppTestBase
         var chainId = "AELF";
         var appGrain = Cluster.Client.GetGrain<IAppSubscriptionGrain>(GrainIdHelper.GenerateAppSubscriptionGrainId(_appInfoProvider.AppId));
         
-        var currentVersion = await appGrain.AddSubscriptionAsync(new SubscriptionManifest
+        var currentVersion = (await appGrain.AddSubscriptionAsync(new SubscriptionManifest
         {
             SubscriptionItems = new List<Subscription>
             {
@@ -163,7 +163,7 @@ public class SubscribedBlockHandlerTests : AeFinderAppTestBase
                     StartBlockNumber = 100000
                 }
             }
-        }, new byte[] { });
+        }, new byte[] { })).NewVersion;
 
         var blockPusherInfoGrain = Cluster.Client.GetGrain<IBlockPusherInfoGrain>(
             GrainIdHelper.GenerateBlockPusherGrainId(_appInfoProvider.AppId, currentVersion, chainId));
@@ -199,7 +199,7 @@ public class SubscribedBlockHandlerTests : AeFinderAppTestBase
         var chainId = "AELF";
         var appGrain = Cluster.Client.GetGrain<IAppSubscriptionGrain>(GrainIdHelper.GenerateAppSubscriptionGrainId(_appInfoProvider.AppId));
         
-        var currentVersion = await appGrain.AddSubscriptionAsync(new SubscriptionManifest
+        var currentVersion = (await appGrain.AddSubscriptionAsync(new SubscriptionManifest
         {
             SubscriptionItems = new List<Subscription>
             {
@@ -209,7 +209,7 @@ public class SubscribedBlockHandlerTests : AeFinderAppTestBase
                     StartBlockNumber = 10000
                 }
             }
-        }, new byte[] { });
+        }, new byte[] { })).NewVersion;
 
         var blockPusherInfoGrain = Cluster.Client.GetGrain<IBlockPusherInfoGrain>(
             GrainIdHelper.GenerateBlockPusherGrainId(_appInfoProvider.AppId, currentVersion, chainId));
