+++ test/AeFinder.Application.Tests/BlockScan/BlockScanAppServiceTests.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using AeFinder.Grains;
 using AeFinder.Grains.Grain.BlockPush;
 using AeFinder.Grains.Grain.Subscriptions;
+using AeFinder.Subscriptions;
 using Orleans;
 using Shouldly;
 using Xunit;
@@ -11,11 +12,13 @@ namespace AeFinder.BlockScan;
 
 public class BlockScanAppServiceTests : AeFinderApplicationOrleansTestBase
 {
-    private IBlockScanAppService _blockScanAppService;
-    private IClusterClient _clusterClient;
+    private readonly IBlockScanAppService _blockScanAppService;
+    private ISubscriptionAppService _subscriptionAppService;
+    private readonly IClusterClient _clusterClient;
 
     public BlockScanAppServiceTests()
     {
+        _subscriptionAppService = GetRequiredService<ISubscriptionAppService>();
         _blockScanAppService = GetRequiredService<IBlockScanAppService>();
         _clusterClient = GetRequiredService<IClusterClient>();
     }
@@ -38,7 +41,7 @@ public class BlockScanAppServiceTests : AeFinderApplicationOrleansTestBase
             }
         };
 
-        var version1 = await _blockScanAppService.AddSubscriptionAsync(appId, subscriptionInput, new byte[1]);
+        var version1 = await _subscriptionAppService.AddSubscriptionAsync(appId, subscriptionInput, new byte[1]);
 
         var subscription = await _blockScanAppService.GetSubscriptionAsync(appId);
         subscription.CurrentVersion.Version.ShouldBe(version1);
@@ -85,7 +88,7 @@ public class BlockScanAppServiceTests : AeFinderApplicationOrleansTestBase
             }
         };
         
-        var version2 = await _blockScanAppService.AddSubscriptionAsync(appId, subscriptionInput2, new byte[1]);
+        var version2 = await _subscriptionAppService.AddSubscriptionAsync(appId, subscriptionInput2, new byte[1]);
         var id2 = GrainIdHelper.GenerateBlockPusherGrainId(appId, version2, chainId);
         
         subscription = await _blockScanAppService.GetSubscriptionAsync(appId);
@@ -131,7 +134,7 @@ public class BlockScanAppServiceTests : AeFinderApplicationOrleansTestBase
             }
         };
         
-        var version3 = await _blockScanAppService.AddSubscriptionAsync(appId, subscriptionInfo3, new byte[1]);
+        var version3 = await _subscriptionAppService.AddSubscriptionAsync(appId, subscriptionInfo3, new byte[1]);
         var id3 = GrainIdHelper.GenerateBlockPusherGrainId(appId, version3, chainId);
 
         subscription = await _blockScanAppService.GetSubscriptionAsync(appId);
@@ -167,147 +170,5 @@ public class BlockScanAppServiceTests : AeFinderApplicationOrleansTestBase
         
         allScanIds = await blockScanManagerGrain.GetAllBlockPusherIdsAsync();
         allScanIds[chainId].ShouldNotContain(id3);
-    }
-
-    [Fact]
-    public async Task UpdateSubscriptionTest()
-    {
-        var subscriptionInfo1 = new SubscriptionManifestDto()
-        {
-            SubscriptionItems=new List<SubscriptionDto>()
-            {
-                new SubscriptionDto()
-                {
-                    ChainId = "AELF",
-                    OnlyConfirmed = false,
-                    StartBlockNumber = 999,
-                    TransactionConditions = new List<TransactionConditionDto>()
-                    {
-                        new TransactionConditionDto()
-                        {
-                            To = "ToAddress",
-                            MethodNames = new List<string>()
-                            {
-                                "Method1",
-                                "Method2"
-                            }
-                        }
-                    },
-                    LogEventConditions = new List<LogEventConditionDto>()
-                    {
-                        new LogEventConditionDto()
-                        {
-                            ContractAddress = "TokenAddress1",
-                            EventNames = new List<string>()
-                            {
-                                "Transfer",
-                                "Burned"
-                            }
-                        }
-                    }
-                }
-            }
-        };
-        var appId = "test-app";
-        var dll = System.Text.Encoding.UTF8.GetBytes("Program codes");
-        var version1 = await _blockScanAppService.AddSubscriptionAsync(appId, subscriptionInfo1, dll);
-        
-        var subscription = await _blockScanAppService.GetSubscriptionAsync(appId);
-        subscription.CurrentVersion.Version.ShouldBe(version1);
-        subscription.CurrentVersion.SubscriptionManifest.SubscriptionItems.Count.ShouldBe(1);
-        subscription.CurrentVersion.SubscriptionManifest.SubscriptionItems[0].StartBlockNumber.ShouldBe(999);
-        subscription.CurrentVersion.SubscriptionManifest.SubscriptionItems[0].LogEventConditions[0].EventNames.Count.ShouldBe(2);
-        subscription.NewVersion.ShouldBeNull();
-
-        var subscriptionInfo2 = new SubscriptionManifestDto()
-        {
-            SubscriptionItems = new List<SubscriptionDto>()
-            {
-                new SubscriptionDto()
-                {
-                    ChainId = "AELF",
-                    OnlyConfirmed = false,
-                    StartBlockNumber = 999,
-                    TransactionConditions = new List<TransactionConditionDto>()
-                    {
-                        new TransactionConditionDto()
-                        {
-                            To = "ToAddress",
-                            MethodNames = new List<string>()
-                            {
-                                "Method1",
-                                "Method2"
-                            }
-                        }
-                    },
-                    LogEventConditions = new List<LogEventConditionDto>()
-                    {
-                        new LogEventConditionDto()
-                        {
-                            ContractAddress = "TokenAddress1",
-                            EventNames = new List<string>()
-                            {
-                                "Transfer",
-                                "Burned",
-                                "Created",
-                                "TransactionChargedFee"
-                            }
-                        }
-                    }
-                }
-            }
-        };
-        await _blockScanAppService.UpdateSubscriptionInfoAsync(appId, version1, subscriptionInfo2);
-        var subscription2 = await _blockScanAppService.GetSubscriptionAsync(appId);
-        subscription2.CurrentVersion.Version.ShouldBe(version1);
-        subscription2.CurrentVersion.SubscriptionManifest.SubscriptionItems.Count.ShouldBe(1);
-        subscription2.CurrentVersion.SubscriptionManifest.SubscriptionItems[0].LogEventConditions[0].EventNames.Count.ShouldBe(4);
-        
-        var subscriptionInfo3 = new SubscriptionManifestDto()
-        {
-            SubscriptionItems = new List<SubscriptionDto>()
-            {
-                new SubscriptionDto()
-                {
-                    ChainId = "AELF",
-                    OnlyConfirmed = false,
-                    StartBlockNumber = 999,
-                    TransactionConditions = new List<TransactionConditionDto>()
-                    {
-                        new TransactionConditionDto()
-                        {
-                            To = "ToAddress",
-                            MethodNames = new List<string>()
-                            {
-                                "Method1",
-                                "Method2",
-                                "Method3"
-                            }
-                        }
-                    },
-                    LogEventConditions = new List<LogEventConditionDto>()
-                    {
-                        new LogEventConditionDto()
-                        {
-                            ContractAddress = "TokenAddress1",
-                            EventNames = new List<string>()
-                            {
-                                "Transfer",
-                                "Burned",
-                                "Created",
-                                "TransactionChargedFee"
-                            }
-                        }
-                    }
-                }
-            }
-        };
-        await _blockScanAppService.UpdateSubscriptionInfoAsync(appId, version1, subscriptionInfo3);
-        var subscription3 = await _blockScanAppService.GetSubscriptionAsync(appId);
-        subscription3.CurrentVersion.Version.ShouldBe(version1);
-        subscription3.CurrentVersion.SubscriptionManifest.SubscriptionItems.Count.ShouldBe(1);
-        subscription3.CurrentVersion.SubscriptionManifest.SubscriptionItems[0].TransactionConditions[0].MethodNames.Count.ShouldBe(3);
-        subscription3.CurrentVersion.SubscriptionManifest.SubscriptionItems[0].LogEventConditions[0].EventNames.Count.ShouldBe(4);
-
     }
 }
